Replace Jansson with YAJL
JSON rendering is a substantial part of the overall run time for certain
workloads. With dnstable 0.7.0, a 'dnstable_dump -j -r' on a ~600 MB
file with the output piped to /dev/null takes 52 seconds to run, and
results in the following 'perf report --sort=dso' output:

    44.64%  libjansson.so.4.7.0
    39.96%  libc-2.19.so
     6.20%  libz.so.1.2.8
     5.03%  libwdns.so.1.0.0
     2.42%  libdnstable.so.0.0.0
     1.45%  libmtbl.so.0.0.0
     0.20%  [kernel.vmlinux]

I.e., most of the run time is spent inside the Jansson library,
rendering JSON. This commit replaces Jansson with YAJL, which is
substantially faster; the same dnstable_dump now takes 28 seconds to
run, and results in the following perf report:

    44.60%  libc-2.19.so
    19.07%  libyajl.so.2.1.0
    12.77%  libdnstable.so.0.0.0
    11.24%  libz.so.1.2.8
     9.15%  libwdns.so.1.0.0
     2.70%  libmtbl.so.0.0.0
     0.29%  [kernel.vmlinux]

YAJL's output is slightly more compact (it uses no whitespace, by
default, while Jansson uses minimal whitespace), and the keys inside the
maps that it generates are in a different order, but, by definition,
this does not matter.

I verified that the dnstable YAJL output is equivalent to the dnstable
Jansson output by running:

    dnstable_dump -j [-r/-d] <FILE> | jq -S -c . | md5sum

on a ~600 MB test file. "jq -S -c ." canonicalizes the input, and I
ended up with identical checksums when swapping between the two JSON
implementations.