entry: dnstable_entry_to_json(): Replace snprintf() for uint64_t formatting
This commit replaces three calls to snprintf() to format individual
integers with calls to a custom function that formats uint64_t's into
their decimal presentation. YAJL's yajl_gen_number() takes a number
that's already been converted to a string, and while YAJL's generator
API does have a yajl_gen_integer() function, its implementation is still
sprintf() based.

This saves us about 8% of the total run time of a 'dnstable_dump -j -r'
with both a ~20 MB file and a ~600 MB file. The 'perf diff' output for
the printf-related C-library functions looks like this:

    # Baseline    Delta  Shared Object         Symbol
    # ........  .......  ....................  ..................................
    #
         7.97%   -3.94%  libc-2.19.so          [.] vfprintf
         0.55%           libc-2.19.so          [.] _IO_vsnprintf
         0.23%           libc-2.19.so          [.] __snprintf
         0.12%   +0.03%  libc-2.19.so          [.] _IO_vsprintf
         0.08%   +0.01%  libc-2.19.so          [.] sprintf
         0.05%   +0.00%  libc-2.19.so          [.] __vsnprintf_chk
         0.01%           libdnstable.so.0.0.0  [.] snprintf@plt
         0.00%   -0.00%  libwdns.so.1.0.0      [.] __vsnprintf_chk@plt

The custom format function is based on djbdns's fmt_ulong(), which is in
the public domain.

(See http://sources.debian.net/src/djbdns/1:1.05-8/fmt_ulong.c/ for a
copy of the original.)