Support YAJL 1
YAJL 1 is close enough to YAJL 2 that we can support it with a few
ifdef's. YAJL 1 support is desireable, since this is the version that
ships with older OSes like Ubuntu 12.04 LTS and CentOS 6. (Debian wheezy
actually ships with YAJL 2.)

First, we need a custom autoconf macro (MY_CHECK_YAJL_1), since,
unfortunately, previous versions of YAJL didn't ship with a .pc file.
This gets chained into the configure script via the PKG_CHECK_MODULES's
action-if-not-found parameter. That is, it will only run if we fail to
find 'yajl >= 2' via pkg-config. The MY_CHECK_YAJL_1 macro will check
for the yajl/yajl_gen.h header as well as the yajl_gen_alloc2 function
in -lyajl, which only exists in YAJL 1.

Then we can conditionalize based on the HAVE_YAJL_1 define in config.h.
If defined, we only need to:

 - Include <yajl/yajl_gen.h> for YAJL 1, <yajl_gen.h> otherwise.

   Normally something like -I/usr/include/yajl would come in via
   "pkg-config --cflags 'yajl >= 2'". Doing it this way avoids having to
   code an autoconf check that finds the precise include path for the
   YAJL 1 headers, while still allowing the YAJL 2 headers to be
   installed into a non-default location.

 - Paper over the 'unsigned int' vs 'size_t' difference for the
   yajl_print_t callback type.

 - Use yajl_gen_alloc2() to initialize our yajl_gen object (YAJL 1),
   otherwise use yajl_gen_alloc() + yajl_gen_config().