dnstable_query_set_filter_parameter(): Add length parameter
Add a 'length' parameter to dnstable_query_set_filter_parameter(), so
that we can validate the size of the passed in parameter, and so that we
can extend this function in the future to support variable length
arguments, if desired. (Currently this must always be 8, since the only
supported filter parameter types are uint64_t's.)

Also do a memcpy() in set_filter_parameter() now that we have an
explicit length, which avoids a few casts and potentially unaligned
accesses.

Also const-ify.