query_iter_next_ip(): Move dnstable_entry_destroy() up
Move the call to dnstable_entry_destroy() up to the beginning of the
'if (rrtype != it->query->rrtype)' block.

The function will never use the 'ent' instance once we know the rrtype
doesn't match, and by bringing this call up we can delete the now
redundant dnstable_entry_destroy() call in the error handling path.