Limit rrtype-based query optimization for rdata name queries.
dnstable encodes rdata for most rrtypes containing hostnames so
that the hostname is at the beginning of the encoding and
immediately followed by the rrtype. This allows rdata name queries
to select a specific rrtype efficiently by appending that rrtype to
the prefix of the underlying mtbl search.

Not all rrtypes are encoded this way, so we restrict the application
of this optimization to rdata queries for which the rrtypes are
encoded accordingly.