query_iter_next_ip(): fix entry skipping logic
When the query_iter_next_ip encounters an encoded
non-IP entry the bytes corresponding to the rrtype of
an encoded IP entry match the IP rrtype, it cannot
yet safely skip ahead to the prefix of the next IP's
address entries.

For example, the encoding of a AAAA rdata entry with
address 2003:0A0A:0100:: begins with the same bytes as
that of an A rdata entry with address 32.3.10.10, but
sorts before such an A rdata entry due to the trailing
zeroes in the IPv6 address.

To handle this case, we look for potential prefixes of
a valid owner name beyond the rrtype bytes, and extend
the search key with these.