Optimize time filtered queries on filesets
When a query with time filters runs against a fileset, we perform the
mtbl query first against a "filter" subset of the fileset. This filter
fileset is constructed so that any merged entry which matches the time
filters must have at least one unmerged component in the filter subset.

For indirect indexed queries, the query against the filter fileset
replaces the RRSET_NAME_FWD or RDATA_NAME_REV index query against
the full fileset, but the subordinate queries are performed against
the full fileset as before.

For direct queries, the query against the filter fileset runs
concurrently with a query against the full fileset. Entries from
the filter fileset are retreived from the full fileset query with
mtbl_iter_seek(), and processed normally from there.