Add tests for time filtered unaggregated queries
The baseline results for unaggregated queries are acquired from a
fileset containing the same data as the test fileset, with an added
entry to each file. This added entry extends the time range of the
file so that it will be included in every time-filtered search, but
named so that it is hidden from the search results.

Repeat all relevant tests in unaggregated mode.